refactor(core): add main package and rename files
rename files in 'genetic' package (i.e., ApplicationRunner.java
became GeneticAlgorithm.java), this is because the algorithm requred
so many inputs and this feels easier to understand. Remove
AminoAcid.java, this was not necessary as it was a duplicate,
so instead of this, add DataPoint.java in main package. DataPoint
class will be called by other packages.